% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveplot.R
\docType{package}
\name{saveplot}
\alias{saveplot}
\alias{saveplot-package}
\title{saveplot: Utility functions for saving high quality plots}
\description{
The saveplot package provides utility functions for saving high quality
plots to svg and png formats for publication in print and online.
}
