% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveplot.R
\name{saveplot}
\alias{saveplot}
\title{Save a plot as PDF}
\usage{
saveplot(plot = plot(1:10), filename = "test", width = 10, height = 10, ...)
}
\arguments{
\item{plot}{A plot object}

\item{filename}{Filename to save to an PDF}

\item{width}{Width of the image in inches}

\item{height}{Height of the image in inches}

\item{...}{other parameters}
}
\value{

}
\description{
Save a plot as PDF
}
\examples{
saveplot(filename = "test.pdf", plot(1:10))
}
\author{
Zhonghui Gai
}
